Various cleanups and structural changes

+++ orca600_api/orca600_registers.h
@@ -9,6 +9,7 @@
 #define ORCA600_REGISTERS_H_
 
 #include "types.h"
+#include "orca600_memory_map.h"
 
 
 /**
@@ -129,7 +130,7 @@ public:
 	 * @brief Write a state id to control register 3, commanding the Orca to change to the selected mode.
 	 */
 	void set_mode (control_register_3_t sid) {
-		write(address, sid);
+		write(address, (uint16_t)sid);
 	}
 
 	/**
